Moved Struct to its own module
