+++ src/element/mod.rs
@@ -13,16 +13,12 @@
 //! [serde-json-value]: https://docs.serde.rs/serde_json/value/enum.Value.html
 
 use crate::element::builders::{ListBuilder, SExpBuilder, StructBuilder};
-use crate::element::iterators::{
-    ElementsIterator, FieldIterator, FieldValuesIterator, IndexVec, SymbolsIterator,
-};
+use crate::element::iterators::{ElementsIterator, SymbolsIterator};
 use crate::element::reader::ElementReader;
 use crate::ion_eq::IonEq;
-use crate::symbol_ref::AsSymbolRef;
 use crate::text::text_formatter::IonValueFormatter;
 use crate::{Decimal, Int, IonResult, IonType, ReaderBuilder, Symbol, Timestamp};
 use num_bigint::BigInt;
-use std::collections::HashMap;
 use std::fmt::{Display, Formatter};
 
 pub mod builders;
@@ -30,7 +26,10 @@ mod element_stream_reader;
 mod iterators;
 pub mod owned;
 pub mod reader;
+mod r#struct;
 pub mod writer;
+
+pub use r#struct::Struct;
 
 /// Behavior that is common to both [SExp] and [Struct].
 pub trait IonSequence {
@@ -147,200 +146,6 @@ impl IonSequence for SExp {
         self.len() == 0
     }
 }
-
-// This collection is broken out into its own type to allow instances of it to be shared with Arc/Rc.
-#[derive(Debug, Clone)]
-struct Fields {
-    // Key/value pairs in the order they were inserted
-    by_index: Vec<(Symbol, Element)>,
-    // Maps symbols to a list of indexes where values may be found in `by_index` above
-    by_name: HashMap<Symbol, IndexVec>,
-}
-
-impl Fields {
-    /// Gets all of the indexes that contain a value associated with the given field name.
-    fn get_indexes<A: AsSymbolRef>(&self, field_name: A) -> Option<&IndexVec> {
-        field_name
-            .as_symbol_ref()
-            .text()
-            .map(|text| {
-                // If the symbol has defined text, look it up by &str
-                self.by_name.get(text)
-            })
-            .unwrap_or_else(|| {
-                // Otherwise, construct a (cheap, stack-allocated) Symbol with unknown text...
-                let symbol = Symbol::unknown_text();
-                // ...and use the unknown text symbol to look up matching field values
-                self.by_name.get(&symbol)
-            })
-    }
-
-    /// Iterates over the values found at the specified indexes.
-    fn get_values_at_indexes<'a>(&'a self, indexes: &'a IndexVec) -> FieldValuesIterator<'a> {
-        FieldValuesIterator {
-            current: 0,
-            indexes: Some(indexes),
-            by_index: &self.by_index,
-        }
-    }
-
-    /// Gets the last value in the Struct that is associated with the specified field name.
-    ///
-    /// Note that the Ion data model views a struct as a bag of (name, value) pairs and does not
-    /// have a notion of field ordering. In most use cases, field names are distinct and the last
-    /// appearance of a field in the struct's serialized form will have been the _only_ appearance.
-    /// If a field name appears more than once, this method makes the arbitrary decision to return
-    /// the value associated with the last appearance. If your application uses structs that repeat
-    /// field names, you are encouraged to use [get_all] instead.
-    fn get_last<A: AsSymbolRef>(&self, field_name: A) -> Option<&Element> {
-        self.get_indexes(field_name)
-            .and_then(|indexes| indexes.last())
-            .and_then(|index| self.by_index.get(*index))
-            .map(|(_name, value)| value)
-    }
-
-    /// Iterates over all of the values associated with the given field name.
-    fn get_all<A: AsSymbolRef>(&self, field_name: A) -> FieldValuesIterator {
-        let indexes = self.get_indexes(field_name);
-        FieldValuesIterator {
-            current: 0,
-            indexes,
-            by_index: &self.by_index,
-        }
-    }
-
-    /// Iterates over all of the (field name, field value) pairs in the struct.
-    fn iter(&self) -> impl Iterator<Item = &(Symbol, Element)> {
-        self.by_index.iter()
-    }
-}
-
-/// An in-memory representation of an Ion Struct
-#[derive(Debug, Clone)]
-pub struct Struct {
-    fields: Fields,
-}
-
-impl Display for Struct {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        let mut ivf = IonValueFormatter { output: f };
-        ivf.format_struct(self).map_err(|_| std::fmt::Error)?;
-        Ok(())
-    }
-}
-
-impl Struct {
-    pub fn builder() -> StructBuilder {
-        StructBuilder::new()
-    }
-
-    pub fn clone_builder(&self) -> StructBuilder {
-        StructBuilder::with_initial_fields(&self.fields.by_index)
-    }
-
-    /// Returns an iterator over the field name/value pairs in this Struct.
-    pub fn fields(&self) -> impl Iterator<Item = (&Symbol, &Element)> {
-        self.fields
-            .iter()
-            // Here we convert from &(name, value) to (&name, &value).
-            // The former makes a stronger assertion about how the data is being stored. We don't
-            // want that to be a mandatory part of the public API.
-            .map(|(name, element)| (name, element))
-    }
-
-    fn fields_eq(&self, other: &Self) -> bool {
-        // For each field name in `self`, get the list of indexes that contain a value with that name.
-        for (field_name, field_value_indexes) in &self.fields.by_name {
-            let other_value_indexes = match other.fields.get_indexes(field_name) {
-                Some(indexes) => indexes,
-                // The other struct doesn't have a field with this name so they're not equal.
-                None => return false,
-            };
-
-            if field_value_indexes.len() != other_value_indexes.len() {
-                // The other struct has fields with the same name, but a different number of them.
-                return false;
-            }
-
-            for field_value in self.fields.get_values_at_indexes(field_value_indexes) {
-                if other
-                    .fields
-                    .get_values_at_indexes(other_value_indexes)
-                    .all(|other_value| !field_value.ion_eq(other_value))
-                {
-                    // Couldn't find an equivalent field in the other struct
-                    return false;
-                }
-            }
-        }
-
-        // If all of the above conditions hold, the two structs are equal.
-        true
-    }
-
-    /// Returns the number of fields in this Struct.
-    pub fn len(&self) -> usize {
-        self.fields.by_index.len()
-    }
-
-    /// Returns `true` if this struct has zero fields.
-    pub fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-}
-
-impl<K, V> FromIterator<(K, V)> for Struct
-where
-    K: Into<Symbol>,
-    V: Into<Element>,
-{
-    /// Returns an owned struct from the given iterator of field names/values.
-    fn from_iter<I: IntoIterator<Item = (K, V)>>(iter: I) -> Self {
-        let mut by_index: Vec<(Symbol, Element)> = Vec::new();
-        let mut by_name: HashMap<Symbol, IndexVec> = HashMap::new();
-        for (field_name, field_value) in iter {
-            let field_name = field_name.into();
-            let field_value = field_value.into();
-
-            by_name
-                .entry(field_name.clone())
-                .or_insert_with(IndexVec::new)
-                .push(by_index.len());
-            by_index.push((field_name, field_value));
-        }
-
-        let fields = Fields { by_index, by_name };
-        Self { fields }
-    }
-}
-
-impl Struct {
-    pub fn iter(&self) -> FieldIterator<'_> {
-        FieldIterator::new(&self.fields.by_index)
-    }
-
-    pub fn get<A: AsSymbolRef>(&self, field_name: A) -> Option<&Element> {
-        self.fields.get_last(field_name)
-    }
-
-    pub fn get_all<A: AsSymbolRef>(&self, field_name: A) -> FieldValuesIterator<'_> {
-        self.fields.get_all(field_name)
-    }
-}
-
-impl PartialEq for Struct {
-    fn eq(&self, other: &Self) -> bool {
-        // check if both fields have same length
-        self.len() == other.len()
-            // we need to test equality in both directions for both fields
-            // A good example for this is annotated vs not annotated values in struct
-            //  { a:4, a:4 } vs. { a:4, a:a::4 } // returns true
-            //  { a:4, a:a::4 } vs. { a:4, a:4 } // returns false
-            && self.fields_eq(other) && other.fields_eq(self)
-    }
-}
-
-impl Eq for Struct {}
 
 impl IonEq for Value {
     fn ion_eq(&self, other: &Self) -> bool {
